Add trivial test on bztree delete

+++ bztree/bztree.h
@@ -307,7 +307,7 @@ class BzTree {
     root = LeafNode::New();
   }
   void Dump();
-  ReturnCode Insert(const char *key, uint64_t key_size, uint64_t payload);
+  ReturnCode Insert(const char *key, uint16_t key_size, uint64_t payload);
   ReturnCode Read(const char *key, uint16_t key_size, uint64_t *payload);
   ReturnCode Update(const char *key, uint16_t key_size, uint64_t payload);
   ReturnCode Upsert(const char *key, uint16_t key_size, uint64_t payload);
